requirements.txt
twilio
colorama
requests
elasticsearch
python-dotenv
